+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 * Fixed an issue where **validate** did not properly check `conf.json` when the latter is modified.
 * Fixed an issue in the **upload** command, where a prompt was not showing on the console.
 * Fixed an issue where running **lint** failed installing dependencies in containers.
+* Added the **pre-commit** command, to improve code quality of XSOAR content.
+* Added the **run-unit-tests** command, to run unit tests of given content items inside their respective docker images.
+* Added support for filepath arguments in the **validate** and **format** commands.
+* Added pre-commit hooks for `validate`, `format`, `run-unit-tests` and `update-docker-image` commands.
 
 ## 1.11.0
 * **Note: Demisto-SDK will soon stop supporting Python 3.8**
@@ -86,7 +90,6 @@
 * Updated the default general `fromVersion` value on **format** to `6.8.0`
 * Fixed an issue where **lint** sometimes failed when using the `-cdam` flag due to wrong file duplications filtering.
 * Added the content graph to **validate**, use with the `--graph` flag.
-
 
 ## 1.9.0
 * Fixed an issue where the Slack notifier was using a deprecated argument.